+++ CHANGES.rst
@@ -103,7 +103,7 @@ v1.0.15
 -----------
 * Fix corrupt release
 
-v1.0.16[unreleased]
+v1.0.16
 -----------
 * Add name filtering to users
 * Deprecate /country-cities/ route and implement /available-cities/ route
@@ -112,3 +112,7 @@ v1.0.16[unreleased]
 * Allow filtering a component by multiple types using an OR filter. (Pass multiple components with different types to use AND filter)
 * Name field becomes EdgeNgram on all indexes
 * On ProjectSearchResource, add setting to include closed projects on query by default
+* Handle dynamic addresses
+
+v1.0.17[unreleased]
+-----------